Update 11.Primality testing prime or not.py

+++ Python/Assignment-1/11.Primality testing prime or not.py
@@ -1,4 +1,4 @@
-number1=float(input("Enter any Number: "))
+number1=int(input("Enter any Number: "))
 number=int(number1)
 if(number>1):
     for i in range(2,(number//2)+1):
